requirements.txt
amazon.ion==0.10.0
Brotli==1.0.9
cbor2==5.4.6
cffi==1.15.1
cryptography==40.0.2
msgpack==1.0.5
protobuf==4.23.2
pycparser==2.21
PyNaCl==1.5.0
securesystemslib==0.28.0
zstandard==0.21.0
